requirements/doc.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
alabaster==0.7.12         # via sphinx
attrs==20.3.0             # via -r requirements/base.in
babel==2.9.0              # via sphinx
bleach==3.2.1             # via readme-renderer
certifi==2020.12.5        # via requests
cffi==1.14.4              # via cryptography
chardet==4.0.0            # via doc8, requests
colorama==0.4.4           # via twine
cryptography==3.3.1       # via secretstorage
django==2.2.17            # via -r requirements/base.in
doc8==0.8.1               # via -r requirements/doc.in
docutils==0.16            # via doc8, readme-renderer, restructuredtext-lint, sphinx
edx-sphinx-theme==1.6.0   # via -r requirements/doc.in
idna==2.10                # via requests
imagesize==1.2.0          # via sphinx
jeepney==0.6.0            # via keyring, secretstorage
jinja2==2.11.2            # via sphinx
keyring==21.7.0           # via twine
markupsafe==1.1.1         # via jinja2
packaging==20.8           # via bleach, sphinx
pbr==5.5.1                # via stevedore
pkginfo==1.6.1            # via twine
pycparser==2.20           # via cffi
pygments==2.7.3           # via doc8, readme-renderer, sphinx
pyparsing==2.4.7          # via packaging
python-dateutil==2.8.1    # via -r requirements/base.in
pytz==2020.4              # via babel, django
readme-renderer==28.0     # via -r requirements/doc.in, twine
requests-toolbelt==0.9.1  # via twine
requests==2.25.1          # via requests-toolbelt, sailthru-client, sphinx, twine
restructuredtext-lint==1.3.2  # via doc8
rfc3986==1.4.0            # via twine
sailthru-client==2.2.3    # via -r requirements/base.in
secretstorage==3.3.0      # via keyring
simplejson==3.17.2        # via sailthru-client
six==1.15.0               # via -r requirements/base.in, bleach, cryptography, doc8, edx-sphinx-theme, python-dateutil, readme-renderer
snowballstemmer==2.0.0    # via sphinx
sphinx==3.4.0             # via -r requirements/doc.in, edx-sphinx-theme
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
sqlparse==0.4.1           # via django
stevedore==3.3.0          # via -r requirements/base.in, doc8
tqdm==4.54.1              # via twine
twine==3.2.0              # via -r requirements/doc.in
urllib3==1.26.2           # via requests
webencodings==0.5.1       # via bleach

# The following packages are considered to be unsafe in a requirements file:
# setuptools
